iova/nattype: fix Clang build errors
- drivers/iommu/iova.c: wrap 'shift' declaration with
  #ifdef CONFIG_ARM64_DMA_IOMMU_ALIGNMENT to avoid unused
  variable error when alignment is disabled. Also cast
  MAX_ALIGN(shift) to unsigned int to silence Clang constant
  conversion errors.

- net/ipv4/netfilter/ipt_NATTYPE.c: ensure NATTYPE target
  compiles by aligning with nf_conn struct. Added Makefile
  entry and config guard so 'nattype_entry' field is visible
  when CONFIG_IP_NF_TARGET_NATTYPE_MODULE is set.

These changes resolve Clang 18 build errors in iova and
NATTYPE modules, ensuring compatibility with current kernel
build.

+++ drivers/iommu/iova.c
@@ -270,7 +270,7 @@ move_left:
 	if (!curr) {
 		if (size_aligned)
 			pad_size = iova_get_pad_size(size, limit_pfn,
-						     MAX_ALIGN(shift));
+				    (unsigned int)MAX_ALIGN(shift));
 		if ((iovad->start_pfn + size + pad_size) > limit_pfn) {
 			spin_unlock_irqrestore(&iovad->iova_rbtree_lock, flags);
 			return -ENOMEM;
